backend/SugangProject/requirements.txt

asgiref==3.2.10

awsebcli==3.19.3

beautifulsoup4==4.9.1

bs4==0.0.1

certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4

cryptography==2.9.2
Django==2.1.5
django-cors-headers==2.1.0


idna==2.10

mysqlclient==2.0.1

pycparser==2.20
PyJWT==1.7.1

pytz==2020.1

requests==2.24.0

six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1

urllib3==1.25.10

